Move some logic to finality_base
