+++ relays/lib-substrate-relay/src/finality/source.rs
@@ -18,10 +18,11 @@
 
 use crate::{
 	finality::{FinalitySyncPipelineAdapter, SubstrateFinalitySyncPipeline},
-	finality_base::engine::Engine,
+	finality_base::{
+		engine::Engine, finality_proofs, SubstrateFinalityProof, SubstrateFinalityProofsStream,
+	},
 };
 
-use crate::finality_base::SubstrateFinalityPipeline;
 use async_std::sync::{Arc, Mutex};
 use async_trait::async_trait;
 use bp_header_chain::FinalityProof;
@@ -29,26 +30,14 @@ use codec::Decode;
 use finality_relay::{SourceClient, SourceClientBase};
 use futures::{
 	select,
-	stream::{try_unfold, unfold, Stream, StreamExt, TryStreamExt},
+	stream::{try_unfold, Stream, StreamExt, TryStreamExt},
 };
 use num_traits::One;
-use relay_substrate_client::{
-	BlockNumberOf, BlockWithJustification, Chain, Client, Error, HeaderOf,
-};
+use relay_substrate_client::{BlockNumberOf, BlockWithJustification, Client, Error, HeaderOf};
 use relay_utils::{relay_loop::Client as RelayClient, UniqueSaturatedInto};
-use std::pin::Pin;
 
 /// Shared updatable reference to the maximal header number that we want to sync from the source.
 pub type RequiredHeaderNumberRef<C> = Arc<Mutex<<C as bp_runtime::Chain>::BlockNumber>>;
-
-/// Substrate finality proofs stream.
-pub type SubstrateFinalityProofsStream<P> =
-	Pin<Box<dyn Stream<Item = SubstrateFinalityProof<P>> + Send>>;
-
-/// Substrate finality proof. Specific to the used `FinalityEngine`.
-pub type SubstrateFinalityProof<P> = <<P as SubstrateFinalityPipeline>::FinalityEngine as Engine<
-	<P as SubstrateFinalityPipeline>::SourceChain,
->>::FinalityProof;
 
 /// Substrate node as finality source.
 pub struct SubstrateFinalitySource<P: SubstrateFinalitySyncPipeline, SourceClnt> {
@@ -216,39 +205,7 @@ impl<P: SubstrateFinalitySyncPipeline, SourceClnt: Client<P::SourceChain>>
 	type FinalityProofsStream = SubstrateFinalityProofsStream<P>;
 
 	async fn finality_proofs(&self) -> Result<Self::FinalityProofsStream, Error> {
-		Ok(unfold(
-			P::FinalityEngine::source_finality_proofs(&self.client).await?,
-			move |mut subscription| async move {
-				loop {
-					let log_error = |err| {
-						log::error!(
-							target: "bridge",
-							"Failed to read justification target from the {} justifications stream: {:?}",
-							P::SourceChain::NAME,
-							err,
-						);
-					};
-
-					let next_justification = subscription.next().await?;
-
-					let decoded_justification =
-						<P::FinalityEngine as Engine<P::SourceChain>>::FinalityProof::decode(
-							&mut &next_justification[..],
-						);
-
-					let justification = match decoded_justification {
-						Ok(j) => j,
-						Err(err) => {
-							log_error(format!("decode failed with error {err:?}"));
-							continue
-						},
-					};
-
-					return Some((justification, subscription))
-				}
-			},
-		)
-		.boxed())
+		finality_proofs::<P>(&self.client).await
 	}
 }
