+++ relays/lib-substrate-relay/src/finality_base/mod.rs
@@ -19,9 +19,12 @@
 
 pub mod engine;
 
+use crate::finality_base::engine::Engine;
 use async_trait::async_trait;
-use relay_substrate_client::{Chain, ChainWithTransactions};
-use std::fmt::Debug;
+use codec::Decode;
+use futures::{stream::unfold, Stream, StreamExt};
+use relay_substrate_client::{Chain, ChainWithTransactions, Client, Error};
+use std::{fmt::Debug, pin::Pin};
 
 /// Substrate -> Substrate finality related pipeline.
 #[async_trait]
@@ -31,5 +34,53 @@ pub trait SubstrateFinalityPipeline: 'static + Clone + Debug + Send + Sync {
 	/// Headers of the `SourceChain` are submitted to this chain.
 	type TargetChain: ChainWithTransactions;
 	/// Finality engine.
-	type FinalityEngine: engine::Engine<Self::SourceChain>;
+	type FinalityEngine: Engine<Self::SourceChain>;
+}
+
+/// Substrate finality proof. Specific to the used `FinalityEngine`.
+pub type SubstrateFinalityProof<P> = <<P as SubstrateFinalityPipeline>::FinalityEngine as Engine<
+	<P as SubstrateFinalityPipeline>::SourceChain,
+>>::FinalityProof;
+
+/// Substrate finality proofs stream.
+pub type SubstrateFinalityProofsStream<P> =
+	Pin<Box<dyn Stream<Item = SubstrateFinalityProof<P>> + Send>>;
+
+/// Subscribe to new finality proofs.
+pub async fn finality_proofs<P: SubstrateFinalityPipeline>(
+	client: &impl Client<P::SourceChain>,
+) -> Result<SubstrateFinalityProofsStream<P>, Error> {
+	Ok(unfold(
+		P::FinalityEngine::source_finality_proofs(client).await?,
+		move |mut subscription| async move {
+			loop {
+				let log_error = |err| {
+					log::error!(
+						target: "bridge",
+						"Failed to read justification target from the {} justifications stream: {:?}",
+						P::SourceChain::NAME,
+						err,
+					);
+				};
+
+				let next_justification = subscription.next().await?;
+
+				let decoded_justification =
+					<P::FinalityEngine as Engine<P::SourceChain>>::FinalityProof::decode(
+						&mut &next_justification[..],
+					);
+
+				let justification = match decoded_justification {
+					Ok(j) => j,
+					Err(err) => {
+						log_error(format!("decode failed with error {err:?}"));
+						continue
+					},
+				};
+
+				return Some((justification, subscription))
+			}
+		},
+	)
+	.boxed())
 }
